requirements.txt
accelerate==1.10.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
bitsandbytes==0.47.0
cd-fvd==0.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
conda-pack==0.8.1
contourpy==1.3.1
controlnet-aux==0.0.9
cupy-cuda117==10.6.0
cycler==0.12.1
datasets==3.4.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.16.3
denku==0.0.51
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.2.2
fastapi==0.115.7
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.19.1
fire==0.5.0
fonttools==4.55.5
frozenlist==1.5.0
fsspec==2025.7.0
fvmd==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.13.0
gradio_client==1.6.0
h11==0.14.0
h5py==3.12.1
hf-xet==1.1.8
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
lightning-utilities==0.14.3
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.60.0
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
optimum==1.24.0
orjson==3.10.15
packaging==25.0
pandas==2.2.3
peft==0.10.0
pillow==11.3.0
platformdirs==4.3.6
prettytable==3.9.0
prodigyopt==1.1.2
proglog==0.1.10
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyav==14.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
qudida==0.0.4
regex==2025.7.34
requests==2.32.5
rich==13.9.4
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scikit-video==1.1.11
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spaces==0.32.0
spandrel==0.4.1
starlette==0.45.2
stringzilla==3.11.3
sympy==1.14.0
tensorboardX==2.6.2.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.6.7
tokenizers==0.21.4
tomlkit==0.13.2
torch==2.8.0
torchaudio==2.1.0+cu121
torchmetrics==1.7.1
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
typer==0.15.1
typing_extensions==4.14.1
tzdata==2025.1
ultralytics==8.3.66
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.0
wandb==0.19.4
wcwidth==0.2.13
websockets==14.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
